third_party/blink/web_tests/fast/dom/custom/registration-context-delete-during-attribute-changed-retrieval-expected.txt

CONSOLE WARNING: line 24: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests destroying a context during registration at the point when the attributeChanged callback is retrieved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS destroyed context
PASS caught expected exception "InvalidStateError: Failed to execute 'registerElement' on 'Document': Registration failed for type 'x-a'. The context is no longer valid."
PASS child done
PASS successfullyParsed is true

TEST COMPLETE
